break-out CollectionsIndexHeader
