+++ src/components/Collections/index.js
@@ -1,60 +1,30 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {
-  StyleSheet, View, SafeAreaView, StatusBar, TouchableOpacity,
+  StyleSheet, View, SafeAreaView,
 } from 'react-native';
 import { shape } from 'prop-types';
-import {
-  Header, Right, Body, Title, Left,
-} from 'native-base';
-import { Feather } from '@expo/vector-icons'; // eslint-disable-line import/no-extraneous-dependencies
 
-import Colors from '../../constants/Colors';
+import CollectionsIndexHeader from './CollectionsIndexHeader';
 import CollectionRow from '../CollectionRow/CollectionRow';
-import CollectionsDialog, { COLLECTIONS_DIALOG_ACTIONS, CollectionsDialogText } from '../Dialog/CollectionsDialog';
 
 const CollectionsIndex = ({
   navigation,
   collections,
-}) => {
-  const [collectionsDialogText, setCollectionsDialogText] = useState(null);
-
-  const handleNewCollectionPress = () => {
-    setCollectionsDialogText(CollectionsDialogText[COLLECTIONS_DIALOG_ACTIONS.CREATE]);
-  };
-
-  return (
-    <SafeAreaView style={styles.safeAreaView}>
-      <StatusBar backgroundColor={Colors.primary} barStyle="dark-content" />
-      <Header style={styles.header}>
-        <Left />
-        <Body>
-          <Title style={styles.headerText}>Collections</Title>
-        </Body>
-        <Right>
-          <TouchableOpacity onPress={handleNewCollectionPress}>
-            <Feather name="plus-square" size={24} color={Colors.headerIcon} />
-          </TouchableOpacity>
-        </Right>
-      </Header>
-      <View style={styles.collectionRowContainer}>
-        {Object.entries(collections).map(([id, { label }]) => (
-          <CollectionRow
-            key={id}
-            collectionId={id}
-            label={label}
-            navigation={navigation}
-          />
-        ))}
-      </View>
-      {collectionsDialogText && (
-        <CollectionsDialog
-          collectionsDialogText={collectionsDialogText}
-          setCollectionsDialogText={setCollectionsDialogText}
+}) => (
+  <SafeAreaView style={styles.safeAreaView}>
+    <CollectionsIndexHeader />
+    <View style={styles.collectionRowContainer}>
+      {Object.entries(collections).map(([id, { label }]) => (
+        <CollectionRow
+          key={id}
+          collectionId={id}
+          label={label}
+          navigation={navigation}
         />
-      )}
-    </SafeAreaView>
-  );
-};
+      ))}
+    </View>
+  </SafeAreaView>
+);
 
 CollectionsIndex.propTypes = {
   navigation: shape({}).isRequired,
@@ -70,13 +40,5 @@ const styles = StyleSheet.create({
   },
   collectionRowContainer: {
     alignItems: 'center',
-  },
-  header: {
-    backgroundColor: Colors.headerBackground,
-    height: 50,
-  },
-  headerText: {
-    color: 'black',
-    fontSize: 18,
   },
 });
